chore: renamed `searchResult.ts` -> `searchResults.ts`
