+++ js/searchResults.ts
